src/requirements.txt
attrs==25.3.0
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.2
cssselect==1.3.0
defusedxml==0.7.1
filelock==3.18.0
greenlet==3.1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.3.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
parsel==1.10.0
Protego==0.4.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
SQLAlchemy==2.0.40
tldextract==5.1.3
Twisted==24.11.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
w3lib==2.3.1
zope.interface==7.2
